requirements.txt
requests>=2.31
beautifulsoup4>=4.12
psycopg2-binary>=2.9
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
lifelines>=0.27
